distgit: Add dist_git_bugs shared function
Planned to be used by automilos and tedude test modules.

Extracts a list of bugs from git commit log diff for a component
in the pipeline.

The method for choosing the baseline build for the diff can be
changed using `baseline-method` option.

Also add some nice test options to make the execution quicker while
developing.

Example

    $ citool brew --name systemd --tag rhel-8.2.0 dist-git --list-bugs --baseline-method previous-build

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>