CI configuration can live in any files really, not just `ci.fmf` (TFT-152).
While thinking about how to detect if CI configuration is present,
I first thought, we will need to clone the repository to find out
(with the help of `tmt plan ls`).

Later I realized that it is a big overkill and the best way (and
easiest) way to do it is just to check for FMF metadata presence
in the dist-git repo - i.e. look if `.fmf/version` file present.

**integration tests**
* :ok: [tmt](https://baseos-stg-jenkins.rhev-ci-vms.eng.rdu2.redhat.com/job/integration-tests/488/artifact/report.html)
* :ok: [sti](https://baseos-stg-jenkins.rhev-ci-vms.eng.rdu2.redhat.com/job/integration-tests/487/artifact/report.html)

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>

Closes TFT-152