test_schedule_tmt: do not use copr repo, install from updates-testing
As discussed with @psplicha, tmt is moving to building the copr repo
automatically after merge to master. This can cause unwanted instability
which we want to mitigate.

From now on, we will consume tmt from fedora updates testing repos,
which provides more stability and at an earlier stage then release
to stable repos.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>