install-mbs-build-execute: support odcs-options-separator
We have a use case where we need to supply to `--odcs-options`
multiple options. Using just space in separating them, i.e.

  --odcs-options="--module-defaults-url=X --module-defaults-commit=Y"

Does not work with Jenkins well, so we are adding a common concept using
a separator instead of the space. So with the default separator the
above option would be:

   --odcs-options="--module-defaults-url=X#-#-#-#-#--module-defaults-commit=Y"

Which is Jenkins friendly and will work.

Just as with other modules, to keep consistent state, we provide an
option to override the default if anybody would need it.

Relates to TFT-304

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>