artemis: add remove_from_list function
When artemis guest is removed not in the destroy phase of the module, artemis module kept the guest object in `guests` list. When the destroy phase is started, artemis tries to delete the deleted guest. The method adds the ability to remove the guest from the list in any time.