beaker: add remove_from_list function
When beaker guest is removed not in the destroy phase of the module, beaker module kept the guest object in `_dynamic_guests` list. When the destroy phase is started, beaker tries to delete the deleted guest. The method adds the ability to remove the guest from the list in any time.