test-scheduler: track duplicate tecs using string as a key
Using the tec itself is almost safe. It is immutable, therefore it can be used as a key, but in the case of exceptions, logging will try to log all locals, and maybe emit the log entry as a JSON, and here is the problem: tec as a key cannot be encoded by JSON encoder. Using its string representation should be as safe as tec itself for keys, but it is better when it comes to logging.