Provisioner modules shall not modify list of guests while iterating in destroy()
With the recent change, guests now removes themselves from the list of guests known to the provisioner, but that conflicts with the fact the provisioner iterates over that list. This may cause troubles, and it may be a cause of `artemis` module not releasing all guests, just the first one.