Drop EPEL repo from our moduleinfo file
Our CI worker container image is based on Fedora. Adding EPEL repository
to it is nto a good idea and lately it also caused problems with some
packages, namely copr-cli.

All pakcages needed from copr-cli are actually around in Fedora
repositories, so there is not need to have it at all.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>